payrollsystem/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djoser==2.2.0
djangorestframework_simplejwt==5.3.0
celery==5.2.7
redis==4.6.0
django_celery_results==2.5.1